version 1.0.0alpha
First version of the site where all the main functionality and layout stuff is implemented. There needs to be some pretty major code cleanup at some point, and there are some features I want to work on. But from here, it should just be filling out content (text, images, videos) on the pages.